Readjusted the Adapter

+++ app/src/main/java/com/sukie2/android/currencyconverter/model/CurrencyData.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class CurrencyResponse(
     @SerializedName("base") val base : String,
     @SerializedName("dates") val dates : String,
-    @SerializedName("rates") val rates : Map<String, Double>
+    @SerializedName("rates") val rates : Map<String, Float>
 )
